its a start

+++ arduino/arduino/communications.h
@@ -13,7 +13,7 @@
 
 // functions
 void initCommunication();
-void sendControl(uint8_t targets[]);
+void sendControl(int targets[]);
 void emergencyBrake();
